requirements.txt
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.26.0
googleapis-common-protos==1.70.0
absl-py==2.3.1
anyio==4.9.0
appdirs==1.4.4
argon2-cffi
arrow
astroid
attrs
autopep8
babel
bcrypt
beautifulsoup4
black
bleach
bokeh
bottleneck
cachetools
certifi
cffi
chardet
charset-normalizer
click
cloudpickle
colorama
cookiecutter
cryptography
cycler
dask
debugpy
decorator
defusedxml
Deprecated
dill
distributed
docutils
et_xmlfile
executing
fastjsonschema
filelock
flake8
Flask
fonttools
frozenlist
fsspec
gitdb
GitPython

grpcio==1.73.1
grpcio-status==1.71.2
h11
h5py
holoviews
httpcore
httplib2==0.22.0
httpx
hvplot
idna
imageio
imagesize
imbalanced-learn
immutabledict==4.2.1
importlib_metadata
iniconfig
intake
ipykernel
ipython
ipywidgets
isort
itsdangerous
jedi
Jinja2
jmespath
joblib
json5
jsonpatch
jsonpointer==2.1
jsonschema
jupyter
jupyter-console
jupyter-client
jupyter-core
jupyter-server
jupyterlab
kiwisolver
lazy_loader
linkify-it-py
llvmlite
lmdb
locket
lxml==5.3.0
Mako
Markdown
markdown-it-py
MarkupSafe
matplotlib==3.10.0
matplotlib-inline
mccabe
mdit-py-plugins
mdurl
mistune
more-itertools
mpmath
msgpack
multidict
multipledispatch
mypy
mypy_extensions
narwhals
numpy
pandas
protobuf
requests
scikit-learn
scipy
seaborn
sqlalchemy
sympy
tabulate
toml
toolz
tornado
tqdm
typing-extensions
urllib3
Werkzeug
widgetsnbextension
zipp
python-dotenv
